+++ tests/frontends/mpd/current_playlist_test.py
@@ -321,6 +321,15 @@ class CurrentPlaylistHandlerTest(unittest.TestCase):
         self.assert_(u'Title: c' in result)
         self.assert_(u'OK' in result)
 
+    def test_plchanges_with_minus_one_returns_entire_playlist(self):
+        self.b.current_playlist.load(
+            [Track(name='a'), Track(name='b'), Track(name='c')])
+        result = self.h.handle_request(u'plchanges "-1"')
+        self.assert_(u'Title: a' in result)
+        self.assert_(u'Title: b' in result)
+        self.assert_(u'Title: c' in result)
+        self.assert_(u'OK' in result)
+
     def test_plchangesposid(self):
         self.b.current_playlist.load([Track(), Track(), Track()])
         result = self.h.handle_request(u'plchangesposid "0"')